Add og:image to Facebook Open Graph API metadata
The Theme Set-up says that this theme uses Facebook's Open Graph API for the feature and thumb images, however, the code is missing the explicitly state the image for Facebook sharing. 

According to Facebook's docs https://developers.facebook.com/docs/sharing/best-practices#tags: "og:image – This is an image associated with your media. We suggest that you use an image of at least 1200x630 pixels." So, the feature image size is the most appropriate for this, the fallback is the thumbnail or site.logo, used same syntax as twitter image.